Curso Springboot
Login de aplicación web a base de datos a través de API REST
Ejecutar en el puerto 8080 y 3306

+++ curso Springboot/src/main/resources/static/js/usuarios.js
@@ -2,39 +2,55 @@
 $(document).ready(function() {
     cargarUsuarios();
   $('#usuarios').DataTable();
+  actualizarEmailDelUsuario();
 });
 
+function actualizarEmailDelUsuario() {
+    document.getElementById('txt-email-usuario').outerHTML = localStorage.email;
+}
+
+
 async function cargarUsuarios() {
- const request = await fetch('api/usuarios', {
+  const request = await fetch('api/usuarios', {
     method: 'GET',
-    headers: {
-      'Accept': 'application/json',
-      'Content-Type': 'application/json'
-    }
+    headers: getHeaders()
   });
   const usuarios = await request.json();
 
 
   let listadoHtml = '';
   for (let usuario of usuarios) {
-    let botonEliminar = '<a href="#" onclick="eliminarUsuario('+ usuario.id +')" class="btn btn-danger btn-circle btn-sm"><i class="fas fa-trash"></i></a><a href="#" class="btn btn-danger btn-circle btn-sm"><i class="fas fa-trash"></i></a>';
-    let usuarioHtml = '<tr><td>'+usuario.id+'</td><td>' + usuario.nombre + ' ' + usuario.apellido + '</td><td>'
-                    + usuario.email+'</td><td>'+usuario.telefono
-                    + '</td><td>' +botonEliminar+ '</td></tr>';
-    listadoHtml += usuarioHtml;
+    let botonEliminar = '<a href="#" onclick="eliminarUsuario(' + usuario.id + ')" class="btn btn-danger btn-circle btn-sm"><i class="fas fa-trash"></i></a>';
 
+    let telefonoTexto = usuario.telefono == null ? '-' : usuario.telefono;
+    let usuarioHtml = '<tr><td>'+usuario.id+'</td><td>' + usuario.nombre + ' ' + usuario.apellido + '</td><td>'
+                    + usuario.email+'</td><td>'+telefonoTexto
+                    + '</td><td>' + botonEliminar + '</td></tr>';
+    listadoHtml += usuarioHtml;
   }
 
 document.querySelector('#usuarios tbody').outerHTML = listadoHtml;
 
 }
 
+function getHeaders() {
+    return {
+     'Accept': 'application/json',
+     'Content-Type': 'application/json',
+     'Authorization': localStorage.token
+   };
+}
+
 async function eliminarUsuario(id) {
-const request = await fetch('api/usuarios/'+ id, {
+
+  if (!confirm('¿Desea eliminar este usuario?')) {
+    return;
+  }
+
+ const request = await fetch('api/usuarios/' + id, {
     method: 'DELETE',
-    headers: {
-      'Accept': 'application/json',
-      'Content-Type': 'application/json'
-    }
+    headers: getHeaders()
   });
+
+  location.reload()
 }
